rewrote the section logic
