+++ src/api/requests.ts
@@ -64,7 +64,7 @@ export const getBoughts = (): Promise<IBoughts> =>{
 export const getDiscounts = (): Promise<IDiscounts[]> => {
   return maxkg.get("discount").json();
 };
-
-export const getPopularGoods = (): Promise<IPopularGood[]> => {
-  return maxkg.get("site/popular?page=1").json();
+ 
+export const getPopularGoods = (page: number): Promise<IPopularGood[]> => {
+  return maxkgz.get(`site/popular?page=${page}`).json();
 };
