+++ src/components/HomeComponents/PopularGoods/PopularGoodsSection/PopularGoodsSection.tsx
@@ -1,0 +1,102 @@
+"use client";
+import React, { useState } from "react";
+import Image from "next/image";
+import {
+  DeliveryIcon,
+  HearthIcon,
+  HearthIconRed,
+} from "../../../../../public/Icons/Icons";
+import { IPopularGood } from "@/types/popularGoods";
+
+interface IPopularGoodsSectionProps {
+  goods: IPopularGood[];
+}
+
+const PopularGoodsSection = ({ goods }: IPopularGoodsSectionProps) => {
+  const imageEmpty = "https://max.kg/images/discount/empty-image.png";
+  const startUrl = "https://max.kg/nal/img/";
+  const initialVisibleItems = 10;
+  const [visibleItems] = useState(initialVisibleItems);
+
+  const [likedItems, setLikedItems] = useState<{ [key: string]: boolean }>({});
+
+
+
+
+  const toggleLike = (id: number) => {
+    setLikedItems((prevState) => ({
+      ...prevState,
+      [id]: !prevState[id],
+    }));
+  };
+
+  function formatNumber(number: number) {
+    return number.toString().replace(/\B(?=(\d{3})+(?!\d))/g, " ");
+  }
+
+
+  return (
+    <section className="goods">
+      <div className="cardContainer container">
+        <div className="cardItemContainer">
+          {goods.slice(0, visibleItems).map((item) => {
+            let imageUrl = imageEmpty;
+
+            if (item.photos && item.photos.length > 0) {
+              const photoUrl = item.photos[0].url_part.startsWith("https://")
+                ? `https://goods-photos.static1-sima-land.com/items/${item.art}/0/280.jpg`
+                : `${startUrl}${item.images[0].id_post}/l_${item.photos[0].url_part}`;
+              imageUrl = photoUrl;
+            }
+
+            return (
+              <div key={item.id} className="cardItem">
+                <Image
+                  className="cardItemImg"
+                  src={imageUrl}
+                  width={230}
+                  height={230}
+                  alt={item.naim}
+                />
+                <div className="cardItemPrices">
+                  <p
+                    className={`cardItemPrice ${
+                      item.old_price !== item.price ? "priceWithOld" : ""
+                    }`}
+                  >
+                    {formatNumber(item.old_price)} с
+                  </p>
+                  {item.old_price !== item.price && (
+                    <>
+                      <p className="cardItemOldPrice">
+                        {formatNumber(item.old_price)} с
+                      </p>
+                    </>
+                  )}
+                </div>
+                <p className="cardItemName">{item.naim}</p>
+                <div className="cardItemDelivery">
+                  <DeliveryIcon />
+                  <p className="cardItemDeliveryTitle">{item.ddos}</p>
+                </div>
+                <div className="cardItemBtns">
+                  <div className="cardItemBtnsContainer">
+                    <button className="cardItemBtnsAddBucket">В корзину</button>
+                    <div
+                      className="hearthIcon"
+                      onClick={() => toggleLike(item.id)}
+                    >
+                      {likedItems[item.id] ? <HearthIconRed /> : <HearthIcon />}
+                    </div>
+                  </div>
+                </div>
+              </div>
+            );
+          })}
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default PopularGoodsSection;
